MultiDvdArchive / LocalizedArchive: Almost done
