+++ sources.py
@@ -320,6 +320,7 @@ SOURCES_REL = [
     Source(src="source/game/jmap/CourseMap.cpp", cc='4201_127', opts=REL_OPTS),
     Source(src="source/game/system/DvdArchive.cpp", cc='4201_127', opts=REL_OPTS),
     Source(src="source/game/system/GhostFile.cpp", cc='4201_127', opts=REL_OPTS),
+    Source(src="source/game/system/LocalizedArchive.cpp", cc='4201_127', opts=REL_OPTS),
     Source(src="source/game/system/MultiDvdArchive.cpp", cc='4201_127', opts=REL_OPTS),
     Source(src="source/game/util/Random.cpp", cc='4201_127', opts=REL_OPTS),
     Source(src="source/game/kart/KartComponent.cpp", cc='4201_127', opts=REL_OPTS),
